moved DashbordProvider to ViewDashboard component
